requirements.txt
streamlit
pandas
geopandas
plotly.express
folium
streamlit_folium
folium.plugins
math